split execution part from executor wrappers
