+++ src/Processors/Executors/Execution/PipelineExecutor.cpp
@@ -14,8 +14,8 @@
 #include <Common/setThreadName.h>
 #include <Common/ThreadGroupSwitcher.h>
 #include <Common/logger_useful.h>
-#include <Processors/Executors/PipelineExecutor.h>
-#include <Processors/Executors/ExecutingGraph.h>
+#include <Processors/Executors/Execution/PipelineExecutor.h>
+#include <Processors/Executors/Execution/ExecutingGraph.h>
 #include <QueryPipeline/printPipeline.h>
 #include <QueryPipeline/ReadProgressCallback.h>
 #include <Processors/ISource.h>
