+++ src/Processors/Executors/Execution/tests/gtest_polling_queue.cpp
@@ -1,4 +1,4 @@
-#include <Processors/Executors/PollingQueue.h>
+#include <Processors/Executors/Execution/PollingQueue.h>
 
 #if defined(OS_LINUX) || defined(OS_DARWIN)
