+++ src/Processors/tests/gtest_exception_on_incorrect_pipeline.cpp
@@ -2,7 +2,7 @@
 
 #include <Processors/Sources/SourceFromSingleChunk.h>
 #include <Processors/Sinks/NullSink.h>
-#include <Processors/Executors/PipelineExecutor.h>
+#include <Processors/Executors/Execution/PipelineExecutor.h>
 
 #include <Columns/ColumnsNumber.h>
 #include <DataTypes/DataTypesNumber.h>
