+++ src/Processors/tests/gtest_scatter_round_robin.cpp
@@ -2,7 +2,7 @@
 
 #include <Columns/ColumnsNumber.h>
 #include <DataTypes/DataTypesNumber.h>
-#include <Processors/Executors/PipelineExecutor.h>
+#include <Processors/Executors/Execution/PipelineExecutor.h>
 #include <Processors/ISink.h>
 #include <Processors/Sources/SourceFromChunks.h>
 #include <Processors/Transforms/ScatterByPartitionTransform.h>
