+++ src/Processors/tests/gtest_write_parquet_page_index.cpp
@@ -7,7 +7,7 @@
 #    include <DataTypes/DataTypesNumber.h>
 #    include <IO/WriteBufferFromFile.h>
 #    include <Processors/Executors/CompletedPipelineExecutor.h>
-#    include <Processors/Executors/PipelineExecutor.h>
+#    include <Processors/Executors/Execution/PipelineExecutor.h>
 #    include <Processors/Formats/Impl/ParquetBlockOutputFormat.h>
 #    include <Processors/ISource.h>
 #    include <Processors/Sources/SourceFromChunks.h>
